Resolve first batch of changes requested by @rubenworks
